requirements.win.txt
Flask==3.0.3
itsdangerous==2.2.0
Werkzeug==3.0.4
Jinja2==3.1.4
python-dotenv==1.0.1
requests==2.32.3

numpy==1.24.3
pandas==2.2.3
Pillow==10.4.0

opencv-python==4.8.1.78
faiss-cpu==1.8.0.post1
retina-face==0.0.13
deepface==0.0.81

tensorflow-intel==2.13.1
keras==2.13.1
tensorboard==2.13.0
h5py==3.10.0

google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
httplib2==0.22.0
gspread==6.1.4

protobuf==4.25.3
typing_extensions==4.5.0
grpcio==1.62.2
